# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ESP32 Weather Map API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Device {
  id: number;
  name: string;
  description?: string | null;
  latitude: number;
  longitude: number;
  isVirtual: boolean;
  isActive: boolean;
  /** Unique hardware ID or slug */
  deviceId: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDeviceBody {
  name: string;
  description?: string | null;
  latitude: number;
  longitude: number;
  isVirtual: boolean;
  deviceId: string;
}

export interface UpdateDeviceBody {
  name?: string;
  description?: string | null;
  latitude?: number;
  longitude?: number;
  isActive?: boolean;
}

export interface SensorReading {
  id: number;
  deviceId: number;
  /** Celsius */
  temperature: number;
  /** Percentage 0-100 */
  humidity: number;
  /** hPa */
  pressure: number;
  /** Feels like temperature */
  heatIndex?: number | null;
  /** km/h */
  windSpeed?: number | null;
  /** Degrees 0-360 */
  windDirection?: number | null;
  uvIndex?: number | null;
  /** Sunny, Cloudy, Rainy, etc. */
  weatherCondition?: string | null;
  recordedAt: string;
}

export interface CreateReadingBody {
  temperature: number;
  humidity: number;
  pressure: number;
  heatIndex?: number | null;
  windSpeed?: number | null;
  windDirection?: number | null;
  uvIndex?: number | null;
  weatherCondition?: string | null;
}

export interface DeviceWithReading {
  device: Device;
  latestReading?: SensorReading | null;
}

export interface DashboardSummary {
  totalDevices: number;
  activeDevices: number;
  virtualDevices: number;
  avgTemperature?: number | null;
  avgHumidity?: number | null;
  avgPressure?: number | null;
  lastUpdated?: string | null;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetDeviceReadingsParams = {
  limit?: number;
};