# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ESP32 Weather Map API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all ESP32 devices
 */
export const ListDevicesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  isVirtual: zod.boolean(),
  isActive: zod.boolean(),
  deviceId: zod.string().describe("Unique hardware ID or slug"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListDevicesResponse = zod.array(ListDevicesResponseItem);

/**
 * @summary Create a new ESP32 device (real or virtual)
 */
export const CreateDeviceBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  isVirtual: zod.boolean(),
  deviceId: zod.string(),
});

/**
 * @summary Get a specific device
 */
export const GetDeviceParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDeviceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  isVirtual: zod.boolean(),
  isActive: zod.boolean(),
  deviceId: zod.string().describe("Unique hardware ID or slug"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update device info
 */
export const UpdateDeviceParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDeviceBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
  isActive: zod.boolean().optional(),
});

export const UpdateDeviceResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  latitude: zod.number(),
  longitude: zod.number(),
  isVirtual: zod.boolean(),
  isActive: zod.boolean(),
  deviceId: zod.string().describe("Unique hardware ID or slug"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a device
 */
export const DeleteDeviceParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get latest readings for a device
 */
export const GetDeviceReadingsParams = zod.object({
  id: zod.coerce.number(),
});

export const getDeviceReadingsQueryLimitDefault = 20;

export const GetDeviceReadingsQueryParams = zod.object({
  limit: zod.coerce.number().default(getDeviceReadingsQueryLimitDefault),
});

export const GetDeviceReadingsResponseItem = zod.object({
  id: zod.number(),
  deviceId: zod.number(),
  temperature: zod.number().describe("Celsius"),
  humidity: zod.number().describe("Percentage 0-100"),
  pressure: zod.number().describe("hPa"),
  heatIndex: zod.number().nullish().describe("Feels like temperature"),
  windSpeed: zod.number().nullish().describe("km\/h"),
  windDirection: zod.number().nullish().describe("Degrees 0-360"),
  uvIndex: zod.number().nullish(),
  weatherCondition: zod
    .string()
    .nullish()
    .describe("Sunny, Cloudy, Rainy, etc."),
  recordedAt: zod.coerce.date(),
});
export const GetDeviceReadingsResponse = zod.array(
  GetDeviceReadingsResponseItem,
);

/**
 * @summary Submit sensor data from an ESP32 device
 */
export const CreateReadingParams = zod.object({
  id: zod.coerce.number(),
});

export const CreateReadingBody = zod.object({
  temperature: zod.number(),
  humidity: zod.number(),
  pressure: zod.number(),
  heatIndex: zod.number().nullish(),
  windSpeed: zod.number().nullish(),
  windDirection: zod.number().nullish(),
  uvIndex: zod.number().nullish(),
  weatherCondition: zod.string().nullish(),
});

/**
 * @summary Get the most recent reading for a device
 */
export const GetDeviceLatestReadingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDeviceLatestReadingResponse = zod.object({
  id: zod.number(),
  deviceId: zod.number(),
  temperature: zod.number().describe("Celsius"),
  humidity: zod.number().describe("Percentage 0-100"),
  pressure: zod.number().describe("hPa"),
  heatIndex: zod.number().nullish().describe("Feels like temperature"),
  windSpeed: zod.number().nullish().describe("km\/h"),
  windDirection: zod.number().nullish().describe("Degrees 0-360"),
  uvIndex: zod.number().nullish(),
  weatherCondition: zod
    .string()
    .nullish()
    .describe("Sunny, Cloudy, Rainy, etc."),
  recordedAt: zod.coerce.date(),
});

/**
 * @summary Generate a random sensor reading for a virtual device
 */
export const SimulateReadingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Dashboard stats - total devices, active, avg temperature, etc.
 */
export const GetDashboardSummaryResponse = zod.object({
  totalDevices: zod.number(),
  activeDevices: zod.number(),
  virtualDevices: zod.number(),
  avgTemperature: zod.number().nullish(),
  avgHumidity: zod.number().nullish(),
  avgPressure: zod.number().nullish(),
  lastUpdated: zod.coerce.date().nullish(),
});

/**
 * @summary Get latest reading for every device (for map display)
 */
export const GetAllLatestReadingsResponseItem = zod.object({
  device: zod.object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string().nullish(),
    latitude: zod.number(),
    longitude: zod.number(),
    isVirtual: zod.boolean(),
    isActive: zod.boolean(),
    deviceId: zod.string().describe("Unique hardware ID or slug"),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  }),
  latestReading: zod
    .object({
      id: zod.number(),
      deviceId: zod.number(),
      temperature: zod.number().describe("Celsius"),
      humidity: zod.number().describe("Percentage 0-100"),
      pressure: zod.number().describe("hPa"),
      heatIndex: zod.number().nullish().describe("Feels like temperature"),
      windSpeed: zod.number().nullish().describe("km\/h"),
      windDirection: zod.number().nullish().describe("Degrees 0-360"),
      uvIndex: zod.number().nullish(),
      weatherCondition: zod
        .string()
        .nullish()
        .describe("Sunny, Cloudy, Rainy, etc."),
      recordedAt: zod.coerce.date(),
    })
    .nullish(),
});
export const GetAllLatestReadingsResponse = zod.array(
  GetAllLatestReadingsResponseItem,
);