# lib/api-zod/src/generated/types/createReadingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ESP32 Weather Map API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReadingBody {
  temperature: number;
  humidity: number;
  pressure: number;
  heatIndex?: number | null;
  windSpeed?: number | null;
  windDirection?: number | null;
  uvIndex?: number | null;
  weatherCondition?: string | null;
}