# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ESP32 Weather Map API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalDevices: number;
  activeDevices: number;
  virtualDevices: number;
  avgTemperature?: number | null;
  avgHumidity?: number | null;
  avgPressure?: number | null;
  lastUpdated?: Date | null;
}