# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ESP32 Weather Map API
 * OpenAPI spec version: 0.1.0
 */

export interface Device {
  id: number;
  name: string;
  description?: string | null;
  latitude: number;
  longitude: number;
  isVirtual: boolean;
  isActive: boolean;
  /** Unique hardware ID or slug */
  deviceId: string;
  createdAt: Date;
  updatedAt: Date;
}