# lib/api-zod/src/generated/types/deviceWithReading.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ESP32 Weather Map API
 * OpenAPI spec version: 0.1.0
 */
import type { Device } from "./device";
import type { SensorReading } from "./sensorReading";

export interface DeviceWithReading {
  device: Device;
  latestReading?: SensorReading | null;
}