# lib/api-zod/src/generated/types/updateDeviceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ESP32 Weather Map API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDeviceBody {
  name?: string;
  description?: string | null;
  latitude?: number;
  longitude?: number;
  isActive?: boolean;
}